requirements.txt
opencv-python
numpy
matplotlib
requests
Pillow
